update: 各コンテンツのViewをComponentとして切り出し
