+++ homete/Resouces/Localizable.xcstrings
@@ -42,7 +42,13 @@
     "しばらくお待ちください" : {
 
     },
+    "タイムライン" : {
+
+    },
     "デバイスの名前を確認してください" : {
+
+    },
+    "テンプレートを設定する" : {
 
     },
     "パートナーを登録して、家事を分担しましょう！" : {
@@ -52,6 +58,9 @@
 
     },
     "プライバシーポリシー" : {
+
+    },
+    "まだアクティビティがありません" : {
 
     },
     "まだパートナーが登録されていません" : {
@@ -82,6 +91,12 @@
     "不明のエラーが発生しました" : {
 
     },
+    "今日の家事リスト" : {
+
+    },
+    "今日の家事を確認して、家事リストを設定しましょう！" : {
+
+    },
     "今月の獲得ポイント" : {
 
     },
@@ -94,7 +109,19 @@
     "同居人同士でこの画面を開いて近づけてください。自動的に登録が始まります。" : {
 
     },
+    "家事がありません" : {
+
+    },
     "家事テンプレート" : {
+
+    },
+    "家事のテンプレートが設定されていません" : {
+
+    },
+    "家事のテンプレートを設定して、家事を分担しましょう！" : {
+
+    },
+    "家事を設定する" : {
 
     },
     "接続エラー" : {
